+++ src/containers/ProDashboard/components/ProTable/ImageExportButton.tsx
@@ -1,11 +1,9 @@
 import { memo, useState } from 'react'
-import { useRouter } from 'next/router'
 import * as echarts from 'echarts/core'
 import { toast } from 'react-hot-toast'
 import { Icon } from '~/components/Icon'
 import { LoadingSpinner } from '~/components/Loaders'
-import { SubscribeModal } from '~/components/Modal/SubscribeModal'
-import { SubscribePlusCard } from '~/components/SubscribeCards/SubscribePlusCard'
+import { SubscribeProModal } from '~/components/SubscribeCards/SubscribeProCard'
 import { useAuthContext } from '~/containers/Subscribtion/auth'
 import { useDarkModeManager } from '~/contexts/LocalStorage'
 import { useIsClient } from '~/hooks'
@@ -30,7 +28,7 @@ export const ImageExportButton = memo(function ImageExportButton({
 	const { loaders } = useAuthContext()
 	const [showSubscribeModal, setShowSubscribeModal] = useState(false)
 	const isClient = useIsClient()
-	const router = useRouter()
+
 	const [isDark] = useDarkModeManager()
 
 	const loading = loaders.userLoading || isSubscriptionLoading || isLoading
@@ -166,11 +164,7 @@ export const ImageExportButton = memo(function ImageExportButton({
 				{isLoading ? <LoadingSpinner size={12} /> : <Icon name="download-paper" height={12} width={12} />}
 				<span>{smol ? '.png' : 'Image'}</span>
 			</button>
-			{isClient && (
-				<SubscribeModal isOpen={showSubscribeModal} onClose={() => setShowSubscribeModal(false)}>
-					<SubscribePlusCard context="modal" returnUrl={router.asPath} />
-				</SubscribeModal>
-			)}
+			{isClient && <SubscribeProModal isOpen={showSubscribeModal} onClose={() => setShowSubscribeModal(false)} />}
 		</>
 	)
 })
