+++ src/containers/Subscribtion/components/SubscriberContent.tsx
@@ -1,8 +1,8 @@
 import * as Ariakit from '@ariakit/react'
 import toast from 'react-hot-toast'
 import { Icon } from '~/components/Icon'
+import { SubscribeAPICard } from '~/components/SubscribeCards/SubscribeAPICard'
 import { SubscribeEnterpriseCard } from '~/components/SubscribeCards/SubscribeEnterpriseCard'
-import { SubscribePlusCard } from '~/components/SubscribeCards/SubscribePlusCard'
 import { SubscribeProCard } from '~/components/SubscribeCards/SubscribeProCard'
 import { Subscription } from '~/hooks/useSubscribe'
 
@@ -55,20 +55,34 @@ export const SubscriberContent = ({
 			<div className="mb-4 flex flex-col items-center">
 				<span className="mb-1 text-xl font-semibold text-white">Change Subscription</span>
 			</div>
-
 			<div className="mb-8 grid grid-cols-1 gap-4 md:grid-cols-3">
-				<SubscribePlusCard
-					context="account"
-					active={isLlamaFeed && subscription?.provider !== 'trial'}
-					onCancelSubscription={isLlamaFeed ? () => handleManageSubscription('llamafeed') : undefined}
-				/>
-				<SubscribeProCard
-					context="account"
-					active={isPro || isLegacy}
-					onCancelSubscription={isPro ? () => handleManageSubscription('api') : undefined}
-					isLegacyActive={isLegacy}
-				/>
-				<SubscribeEnterpriseCard active={subscription?.type === 'enterprise'} />
+				<div
+					className={`relative flex w-full shrink-0 snap-center flex-col overflow-hidden rounded-xl border border-[#4a4a50] bg-[#22242930] px-4 py-8 shadow-md backdrop-blur-md transition-all duration-300 not-first:hover:transform md:w-auto md:max-w-[400px] md:flex-1 md:shrink md:snap-none md:px-5 md:hover:scale-[1.02]`}
+				>
+					<SubscribeProCard
+						context="account"
+						active={isLlamaFeed && subscription?.provider !== 'trial'}
+						onCancelSubscription={isLlamaFeed ? () => handleManageSubscription('llamafeed') : undefined}
+					/>
+				</div>
+				<div
+					className={`relative flex w-full shrink-0 snap-center flex-col overflow-hidden rounded-xl border border-[#4a4a50] bg-[#22242930] px-4 py-8 shadow-md backdrop-blur-md transition-all duration-300 not-first:hover:transform md:w-auto md:max-w-[400px] md:flex-1 md:shrink md:snap-none md:px-5 md:hover:scale-[1.02]`}
+				>
+					<SubscribeAPICard
+						context="account"
+						active={isPro || isLegacy}
+						onCancelSubscription={isPro ? () => handleManageSubscription('api') : undefined}
+						isLegacyActive={isLegacy}
+					/>
+				</div>
+				<div
+					className={`relative flex w-full shrink-0 snap-center flex-col overflow-hidden rounded-xl border border-[#4a4a50] bg-[#22242930] px-4 py-8 shadow-md backdrop-blur-md transition-all duration-300 not-first:hover:transform md:w-auto md:max-w-[400px] md:flex-1 md:shrink md:snap-none md:px-5 md:hover:scale-[1.02]`}
+				>
+					<h2 className="text-center text-[2rem] font-extrabold whitespace-nowrap">Enterprise</h2>
+					<span className="h-8"></span>
+					<span className="h-7"></span>
+					<SubscribeEnterpriseCard active={subscription?.type === 'enterprise'} />
+				</div>
 			</div>
 
 			{(isPro || isLegacy) && (
@@ -387,8 +401,8 @@ export const SubscriberContent = ({
 										<div className="flex-1">
 											<h4 className="mb-1 font-medium">Enable Overage</h4>
 											<p className="text-sm text-[#8a8c90]">
-												Allow your API calls to continue beyond the 1M monthly limit. Additional usage will be charged at
-												$0.60 per 1,000 calls.
+												Allow your API calls to continue beyond the 1M monthly limit. Additional usage will be charged
+												at $0.60 per 1,000 calls.
 											</p>
 										</div>
 									</div>
